+++ clients/core/src/Application/ApplicationConfiguration/ApplicationQuestionConfig/components/MultiSelectConfig.tsx
