+++ clients/core/src/Application/ApplicationConfiguration/ApplicationQuestionConfig/components/RequiredLabel.tsx
