+++ clients/core/src/Application/ApplicationConfiguration/ApplicationQuestionConfig/components/TextConfig.tsx
